+++ CHANGELOG.md
@@ -25,7 +25,8 @@ This file is used to list changes made in each version of the habitat cookbook.
 - Allow any channel in a service [#120](https://github.com/chef-cookbooks/habitat/pull/120) ([jsirex](https://github.com/jsirex))
 - fix hashbang in expeditor update script [#118](https://github.com/chef-cookbooks/habitat/pull/118) ([joshbrand](https://github.com/joshbrand))
 
-<!-- latest_release unreleased -->
+<!-- latest_release -->
+
 ## Unreleased
 
 ## 0.59.0 (2018-07-17)